+++ infra/resources/_modules/key_vaults/outputs.tf
@@ -1,3 +1,6 @@
-output "kv_name" {
-  value = azurerm_key_vault.com.name
+output "com" {
+  value = {
+    name = azurerm_key_vault.com.name
+    id   = azurerm_key_vault.com.id
+  }
 }